requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
certifi==2024.7.4
cffi==2.0.0
charset-normalizer==3.3.2
cryptography==46.0.3
dj-database-url==2.3.0
Django==5.1
django-cors-headers==4.3.0
django-heroku==0.3.1
django_csp==3.8
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.7
packaging==24.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
whitenoise==6.8.2
